Move test fixture to appropriate app
